recommended fixation

+++ models/purchaseRequest.js
@@ -23,11 +23,11 @@ const PurchaseRequest = sequelize.define(
       defaultValue: "pending",
     },
     amount: {
-      type: DataTypes.DECIMAL(10, 2),
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
      approvedAmount: {
-      type: DataTypes.DECIMAL(10, 2),
+      type: DataTypes.INTEGER,
       allowNull: true,
     },
     requestDate: {
